[settings] refactor(core): migrate to feature-driven structure (wave-1) (#142)
